Основной сервис v2.0 (добавил логики, исправил недочеты)

+++ main-service/src/main/java/ewm/category/service/CategoryServiceImpl.java
@@ -5,7 +5,9 @@ import ewm.category.dto.NewCategoryDto;
 import ewm.category.mapper.CategoryMapper;
 import ewm.category.model.Category;
 import ewm.category.repository.CategoryRepository;
+import ewm.event.repository.EventRepository;
 import ewm.exception.EntityNotFoundException;
+import ewm.exception.ValidationException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -18,6 +20,7 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     public List<CategoryDto> getAll(Integer from, Integer size) {
         Pageable pageable = PageRequest.of(from, size);
@@ -38,6 +41,15 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     public void deleteCategory(Long id) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(Category.class, "Категория с id=" + id + " не найдена"));
+
+        boolean hasEvents = eventRepository.existsByCategoryId(id);
+        if (hasEvents) {
+            throw new ValidationException(Category.class, "Невозможно удалить категорию: к ней привязаны события");
+        }
+
+        categoryRepository.delete(category);
         categoryRepository.deleteById(id);
     }
